Add new app states for loading assets
